SUIT-9626 Fix UT failing after merge with master. Fixed some kind of cyclic module referencing: testLaunchHelper.js <-> requests.js
