+++ lib/testLauncher/SuitestLauncher.js
@@ -9,9 +9,11 @@ const {
 	followChildProcess,
 	validateInput,
 	writeLogs,
+} = require('../utils/testLauncherHelper');
+const {
 	getDevicesDetails,
 	getDeviceName,
-} = require('../utils/testLauncherHelper');
+} = require('../utils/getDeviceInfo');
 const envVars = require('../constants/enviroment');
 const {openSessionUnchained} = require('../commands/openSession');
 const {closeSessionUnchained} = require('../commands/closeSession');
