+++ lib/texts.js
@@ -26,7 +26,7 @@ module.exports = {
 	valueMustBeRepo: () => 'value is expected to be VALUE.REPO when matching against repository element',
 	recursionDetected: () => 'Recursive structure detected',
 
-	unknownError: () => 'Unknown error occurred. If you keep getting this error please get in touch with support@suite.st. If you haven\'t turned off automatic error reporting we are already working hard to fix it.',
+	unknownError: () => 'Unknown error occurred. If you keep getting this error please get in touch with support@suite.st. If you haven\'t turned off automatic error reporting (the sentry option in config) we are already working hard to fix the issue.',
 	// Auth errors
 	authNotAllowed: commandName => `You're not allowed to execute .${commandName} function. Make sure you're logged in with the correct credentials.`,
 	authFailed: () => 'Authentication failed. Make sure you\'re trying to login with correct credentials.',
