+++ lib/utils/testLauncherHelper.js
@@ -8,11 +8,6 @@ const {stColors} = require('../testLauncher/launcherLogger');
 const {snippets: log, launcherLogger} = require('../testLauncher/launcherLogger');
 const texts = require('../texts');
 const path = require('path');
-const request = require('../api/request');
-const endpoints = require('../api/endpoints');
-const {authContext} = require('../context');
-const fetch = require('node-fetch');
-const {filter} = require('ramda');
 
 /**
  * Handle test launcher command error
@@ -158,42 +153,6 @@ function writeLogs(deviceId, userArgs, logDir, deviceName) {
 	return writeStream;
 }
 
-/**
- * Return devices details
- * @param {array} devicesIds array of devices id's
- * @returns {object}
- */
-async function getDevicesDetails(devicesIds) {
-	const authorizeHttp = await authContext.authorizeHttp(endpoints.devices, {
-		method: 'GET',
-	});
-	let response = await request(
-		[endpoints.devices, null, {limit: 100}],
-		authorizeHttp
-	);
-	let devices = [...response.values];
-
-	while (response.next) {
-		response = await fetch(response.next, authorizeHttp);
-		response = await response.json();
-		devices = [...devices, ...response.values];
-	}
-
-	return filter(d => devicesIds.includes(d.deviceId), devices);
-}
-
-/**
- * Returns device name for logs based on device information
- * @param {object} deviceInfo
- */
-function getDeviceName(deviceInfo) {
-	if (deviceInfo.manufacturer || deviceInfo.model) {
-		return `${deviceInfo.manufacturer} - ${deviceInfo.model}`;
-	}
-
-	return '';
-}
-
 module.exports = {
 	handleLauncherError,
 	handleChildResult,
@@ -203,6 +162,4 @@ module.exports = {
 	followChildProcess: followChildProcess,
 	validateInput,
 	writeLogs,
-	getDevicesDetails,
-	getDeviceName,
 };
